+++ coffeeshop/src/main/java/com/coffeeshop/Services/CustomerService.java
@@ -29,7 +29,7 @@ public interface CustomerService {
 
     public OrderEntity getOrderDetails(long orderid);
 
-    public List<OrderEntity> getQueueOrders(int queueid, long shopid);
+    public List<OrderEntity> getQueueOrders(long queueid, long shopid);
 
-    public int waitingCount(int queueid, long shopid);
+    public int waitingCount(long queueid, long shopid);
 }
